Add 'Contracts' tag to Swagger documentation (#657)
A custom document filter has been implemented in Swagger to add a new 'Contracts' tag. The addition includes an description of 'Contracts'. Annotations are included in the SwaggerGen to enable this filter, and test cases have been added to confirm the presence of the new tag in all API version documentations.